requirements.txt
numpy
pandas
torch
torchvision  # (Opcional, caso use imagens)
transformers
seqeval
sentencepiece
matplotlib
pytorch-crf